// src/types/StopPoint.ts
export interface components {
  schemas: {
// ...
    "Tfl-8": {
      modeName: string;
      lineIdentifier: string[];
    };
// ...
    "Tfl-11": {
      naptanId: string;
      platformName: string;
      /** @description The indicator of the stop point e.g. "Stop K" */
      indicator: string;
      /** @description The stop letter, if it could be cleansed from the Indicator e.g. "K" */
      stopLetter: string;
      modes: string[];
      icsCode: string;
      smsCode: string;
      stopType: string;
      stationNaptan: string;
      accessibilitySummary: string;
      hubNaptanCode: string;
      lines: components["schemas"]["Tfl-6"][];
      lineGroup: components["schemas"]["Tfl-7"][];
      lineModeGroups: components["schemas"]["Tfl-8"][];
      fullName: string;
      naptanMode: string;
      status: boolean;
      /** @description A unique identifier. */
      id: string;
      /** @description The unique location of this resource. */
      url: string;
      /** @description A human readable name. */
      commonName: string;
      /**
       * Format: double
       * @description The distance of the place from its search point, if this is the result
       *             of a geographical search, otherwise zero.
       */
      distance: number;
      /** @description The type of Place. See /Place/Meta/placeTypes for possible values. */
      placeType: string;
      /** @description A bag of additional key/value pairs with extra information about this place. */
      additionalProperties: components["schemas"]["Tfl-9"][];
      children: components["schemas"]["Tfl-10"][];
      childrenUrls: string[];
      /**
       * Format: double
       * @description WGS84 latitude of the location.
       */
      lat: number;
      /**
       * Format: double
       * @description WGS84 longitude of the location.
       */
      lon: number;
    };
// ...
    "Tfl-20": {
      /** @description The centre latitude/longitude of this list of StopPoints */
      centrePoint: number[];
      /** @description Collection of stop points */
      stopPoints: components["schemas"]["Tfl-11"][];
      /**
       * Format: int32
       * @description The maximum size of the page in this response i.e. the maximum number of StopPoints
       */
      pageSize: number;
      /**
       * Format: int32
       * @description The total number of StopPoints available across all pages
       */
      total: number;
      /**
       * Format: int32
       * @description The index of this page
       */
      page: number;
    };
// ...
  };
}